+++ database/migrations/2021_11_25_141013_add_last_month_modulus.php
@@ -13,9 +13,11 @@ class AddLastMonthModulus extends Migration
      */
     public function up()
     {
-        Schema::table('billing', function (Blueprint $table) {
-            $table->float('sisa')->nullable();
-        });
+        if (!Schema::hasColumn('billing', 'sisa')) {
+            Schema::table('billing', function (Blueprint $table) {
+                $table->float('sisa')->nullable();
+            });
+        }
     }
 
     /**
@@ -25,8 +27,10 @@ class AddLastMonthModulus extends Migration
      */
     public function down()
     {
-        Schema::table('billing', function (Blueprint $table) {
-            $table->dropColumn('sisa');
-        });
+        if (Schema::hasColumn('billing', 'sisa')) {
+            Schema::table('billing', function (Blueprint $table) {
+                $table->dropColumn('sisa');
+            });
+        }
     }
 }
